refactor(future-cli): a bit more refactoring , for better handling of params & imports
